AWShealthCoach/requirements.txt
# Twilio SDK for sending and receiving messages
twilio>=7.0.0

# OpenAI SDK for interacting with GPT models
openai>=0.22.0

# Python-dotenv for environment variable management
python-dotenv>=0.19.0

# AWS SDK for Python, for interacting with AWS services like DynamoDB and Lambda
boto3>=1.20.0

# Requests, if you're planning to make HTTP requests for testing
requests>=2.25.1
